requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
requests==2.32.5
qwen-agent==0.0.34
ddgs==9.11.1
beautifulsoup4==4.14.3
Pillow==11.3.0
numpy==2.3.3
soundfile==0.13.1
python-dateutil==2.9.0.post0
